EN-997 Extract loggers in a new package
